Fixed scroll configuration issue
If I have global scroll set to true and my tipSettings scroll set to
false the test ( true || false) evaluates to true, when what I wanted
was that particular tip to not scroll. Now the setting on the tip
overloads the global.

There is probably a nicer way of doing this :)